move game script to separate file, add scss styles to index.js
