+++ src/pages/index.html
@@ -4,7 +4,6 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width" />
     <title>Memorygame</title>
-    <link rel="stylesheet" type="text/css" href="./assets/style/main.css" />
   </head>
 
   <body>
